Change to double amount

+++ Sources/CedroStreamingSocket/Models/AggregatedBook/AggregatedBookOffersUpdate.swift
@@ -9,7 +9,7 @@ public struct AggregatedBookOffersUpdate: ServiceModelProtocol {
     /// Preço da oferta
     public var price: Double
     /// Quantidade da oferta
-    public var amount: Int
+    public var amount: Double
     /// Número de ofertas agregadas.
     public var offerNumbers: Int
     /// Data e hora da oferta
@@ -69,9 +69,9 @@ extension AggregatedBookOffersUpdate: ServiceProtocol {
         return price
     }
     
-    private static func decodeAmount(from components: [String]) throws -> Int {
+    private static func decodeAmount(from components: [String]) throws -> Double {
         guard components.indices.contains(6) else { throw CedroServiceError.dontContainsAmount }
-        guard let amount = components[6].intValue else { throw CedroServiceError.invalidAmount }
+        guard let amount = components[6].doubleValue else { throw CedroServiceError.invalidAmount }
         return amount
     }
