+++ Sources/CedroStreamingSocket/Models/BookQuote/BookQuoteOffersUpdate.swift
@@ -11,7 +11,7 @@ public struct BookQuoteOffersUpdate: ServiceModelProtocol {
     /// Preço da oferta
     public var price: Double
     /// Quantidade da oferta
-    public var amount: Int
+    public var amount: Double
     /// Código de identificação da corretora detentora da oferta
     public var broker: Int
     /// Data e hora da oferta
@@ -84,9 +84,9 @@ extension BookQuoteOffersUpdate: ServiceProtocol {
         return price
     }
     
-    private static func decodeAmount(from components: [String]) throws -> Int {
+    private static func decodeAmount(from components: [String]) throws -> Double {
         guard components.indices.contains(7) else { throw CedroServiceError.dontContainsAmount }
-        guard let amount = components[7].intValue else { throw CedroServiceError.invalidAmount }
+        guard let amount = components[7].doubleValue else { throw CedroServiceError.invalidAmount }
         return amount
     }
